backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
plaid-python==27.0.0
anthropic==0.42.0
twilio==9.4.3
httpx==0.28.1
apscheduler==3.11.0
python-dotenv==1.0.1
supabase==2.11.0
PyJWT[crypto]==2.10.1
resend==2.23.0
slowapi==0.1.9
